+++ light.go
@@ -1,16 +1,16 @@
 package launchpad
 
-func (l *launchpad) Light(x, y, g, r uint8) error {
+func (l *launchpad) Light(x, y, g, r int) error {
 	note := x + 16*y
 	velocity := 16*g + r + 8 + 4
 	if y >= 8 {
 		note = x + 104
 
-		_, err := l.write([]byte{0xB0, note, velocity})
+		_, err := l.write([]byte{0xB0, byte(note), byte(velocity)})
 		return err
 	}
 
-	_, err := l.write([]byte{0x90, note, velocity})
+	_, err := l.write([]byte{0x90, byte(note), byte(velocity)})
 	return err
 }
